bug fixes for checkout flow along with test cases
