+++ apps/ex_shop/web/models/line_item.ex
@@ -63,7 +63,7 @@ defmodule ExShop.LineItem do
     Repo.preload(line_item, [:product, :order])
   end
 
-  defp validate_product_availability(%ExShop.LineItem{} = line_item) do
+  def validate_product_availability(%ExShop.LineItem{} = line_item) do
     quantity = line_item.quantity
     # have to make sure product is preloaded
     product_quantity = line_item.product.quantity
@@ -74,7 +74,7 @@ defmodule ExShop.LineItem do
     end
   end
 
-  defp validate_product_availability(model) do
+  def validate_product_availability(model) do
     quantity = get_field(model, :quantity)
     # have to make sure product is preloaded
     product_quantity = get_field(model, :product).quantity
